[Refactor](External) Move Common ODBC Methods to JDBC Class and Add Default config to Disable ODBC Creation
