+++ fe/fe-core/src/main/java/org/apache/doris/analysis/CreateResourceStmt.java
@@ -20,6 +20,7 @@ package org.apache.doris.analysis;
 import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.Resource.ResourceType;
 import org.apache.doris.common.AnalysisException;
+import org.apache.doris.common.Config;
 import org.apache.doris.common.ErrorCode;
 import org.apache.doris.common.ErrorReport;
 import org.apache.doris.common.FeNameFormat;
@@ -92,6 +93,10 @@ public class CreateResourceStmt extends DdlStmt {
         }
         if (resourceType == ResourceType.SPARK && !isExternal) {
             throw new AnalysisException("Spark is external resource");
+        }
+        if (resourceType == ResourceType.ODBC_CATALOG && !Config.enable_odbc_table) {
+            throw new AnalysisException("ODBC table is deprecated, use JDBC instead. Or you can set "
+                + "`enable_odbc_table=true` in fe.conf to enable ODBC again.");
         }
     }
