+++ fe/fe-core/src/test/java/org/apache/doris/analysis/CreateResourceStmtTest.java
@@ -21,6 +21,7 @@ import org.apache.doris.catalog.Env;
 import org.apache.doris.catalog.Resource;
 import org.apache.doris.catalog.Resource.ResourceType;
 import org.apache.doris.common.AnalysisException;
+import org.apache.doris.common.Config;
 import org.apache.doris.common.UserException;
 import org.apache.doris.mysql.privilege.AccessControllerManager;
 import org.apache.doris.mysql.privilege.PrivPredicate;
@@ -73,6 +74,7 @@ public class CreateResourceStmtTest {
         properties = Maps.newHashMap();
         properties.put("type", "odbc_catalog");
         stmt = new CreateResourceStmt(true, false, resourceName2, properties);
+        Config.enable_odbc_table = true;
         stmt.analyze(analyzer);
         Assert.assertEquals(resourceName2, stmt.getResourceName());
         Assert.assertEquals(Resource.ResourceType.ODBC_CATALOG, stmt.getResourceType());
