+++ fe/fe-core/src/test/java/org/apache/doris/catalog/EnvOperationTest.java
@@ -23,6 +23,7 @@ import org.apache.doris.analysis.AlterTableStmt;
 import org.apache.doris.analysis.CreateDbStmt;
 import org.apache.doris.analysis.CreateResourceStmt;
 import org.apache.doris.analysis.CreateTableStmt;
+import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.FeConstants;
 import org.apache.doris.qe.ConnectContext;
@@ -100,6 +101,7 @@ public class EnvOperationTest {
     }
 
     private static void createResource(String sql) throws Exception {
+        Config.enable_odbc_table = true;
         CreateResourceStmt createResourceStmt = (CreateResourceStmt) UtFrameUtils.parseAndAnalyzeStmt(sql, connectContext);
         Env.getCurrentEnv().getResourceMgr().createResource(createResourceStmt);
     }
