+++ fe/fe-core/src/test/java/org/apache/doris/catalog/OdbcCatalogResourceTest.java
@@ -20,6 +20,7 @@ package org.apache.doris.catalog;
 import org.apache.doris.analysis.AccessTestUtil;
 import org.apache.doris.analysis.Analyzer;
 import org.apache.doris.analysis.CreateResourceStmt;
+import org.apache.doris.common.Config;
 import org.apache.doris.common.FeMetaVersion;
 import org.apache.doris.common.UserException;
 import org.apache.doris.common.proc.BaseProcResult;
@@ -86,6 +87,7 @@ public class OdbcCatalogResourceTest {
 
         // host: 127.0.0.1, port: 7777, without driver and odbc_type
         CreateResourceStmt stmt = new CreateResourceStmt(true, false, name, properties);
+        Config.enable_odbc_table = true;
         stmt.analyze(analyzer);
         OdbcCatalogResource resource = (OdbcCatalogResource) Resource.fromStmt(stmt);
         Assert.assertEquals(name, resource.getName());
